fix permisisons checking

+++ commands/moderation/purge.js
@@ -33,12 +33,14 @@ module.exports = class PurgeCommand extends commando.Command {
   }
 
   //eslint-disable-next-line class-methods-use-this
-  hasPermission(msg) {
-    return msg.member.hasPermission('MANAGE_MESSAGES');
-  }
-
-  //eslint-disable-next-line class-methods-use-this
   async run(message, args) {
+    let modrole = message.guild.settings.get('modrole')
+    let adminrole = message.guild.settings.get('adminrole')
+    let modlog = message.guild.settings.get('modlog')
+    if (!modrole || !adminrole || !modlog) return message.reply(`This command is not set up to work! Have someone run \`${message.guild.commandPrefix}settings\` to add the \`mod\`, \`admin\`, and \`modlog\` settings.`)
+    if (!message.member.roles.has(modrole)) {
+      if (!message.member.roles.has(adminrole)) return message.reply(`You do not have permission to do this! Only people with this role can access this command! \`Role Required: ${message.guild.roles.get('modrole')}\`, this is changeable with \`${message.guild.commandPrefix}set add mod @role\``)
+    }
     message.channel.send('PURGING')
     message.channel.bulkDelete(args.toPurge)
       .then(() => {
